Update DeleteContacts.php
Added the ability to delete a specific contact.

+++ LAMPAPI/DeleteContacts.php
@@ -3,6 +3,7 @@
 
 	// Extract variables from the request payload
 	$id = $inData["ID"];
+	$userid = $inData["userID"];
 
 	// Initialize database connection
 	$conn = new mysqli("localhost", "APIMan", "WeLoveAPIMan", "COP4331"); 	
@@ -15,10 +16,10 @@
 	else
 	{
 		// Prepare SQL statement for inserting data
-		$stmt = $conn->prepare("DELETE FROM Contacts WHERE ID = ?");
+		$stmt = $conn->prepare("DELETE FROM Contacts WHERE (ID = ? and userID = ?)");
 
 		// Bind the variables to the SQL statement
-		$stmt->bind_param("s", $id);
+		$stmt->bind_param("ss", $id, $userid);
 
 		// Execute the SQL statement
 		$stmt->execute();
